start working on a from-scratch write of an init script
LSB is a tragic joke that the major players (Red Hat, Debian, Ubuntu)
all claim to be compliant with yet stray strongly while still calling it
"lsb" especially with respect to lsbinit.

This is the start of an effort to see if it's at all possible to get a
cross-linux-distro-sane init script to reduce the maintenance surface
here.

Further, the previous init script here originated from
/etc/init.d/skeleton on Ubuntu which originates from the sysvinit
project which is GPl3 and would likely taint this project by its
presence. Instead, this is a from-scratch implementation of the LSB init
script interface (start, stop, etc) that is unburdened by GPL3
licensing.