Don't break systems on apt-get upgrade
postrm is also executed on upgrades, this broke upgrades if the user was
currently in use, leaving apt in a broken state. After the upgrade the user
would be recreated, but might get a different uid which breaks permissions of
the packaged application.

This patches prevents user deletion unless the package is purged.