PleaseRun::Detector.execute now returns boolean for the 2nd value.
This change helps when an execution fails because of a path or
permissions error. Ruby doesn't expose a constructor for
Process::Status, so we cannot fake one of these. This makes me use
a boolean instead. Oh well :)

Now when a program doesn't exist (like systemctl), the detector
will indicate that a given platform is not present on the current
system.