Don't need to use chroot cmd to change user or chroot since newer Upstarts support it.
Also: Remove log outputting. This didn't work if you set a non-root user in all cases.
The problem was that a non-root user may not be able to create the log
file output. Upstart has not-so-great log configuration options, so we
may have to try and fix this later :\
(https://github.com/elastic/logstash/issues/6198)